Use TinyGo __app_ram_base
This makes sure the SoftDevice start address is configured in only one
place, and thus avoid potential issues when it is changed. It also will
allow setting a different SoftDevice RAM size with a linker flag, in
LLVM 11 (when expressions are allowed in the MEMORY part of a linker
script).