gatts/sd: [WIP] allow placing attribute values in user RAM
WIP: still need to find a platform-compatible way of determining whether
an attribute value must be located in stack or in user RAM for SoftDevice
targets.

Currently, the algorithm works this way: if a characteristics' config
has an attribute value with length greater than zero, the value is set
to be placed in user RAM; otherwise, in stack.

This is a breaking change, since previous software written with non-zero
lengths attrib. values will not be aware that using user RAM instead of
the stack forces them to lock that RAM, i.e., the memory must be reserved
for the lifetime of the attribute. There should be an option for specifying
what attribute will go to RAM and what will go to stack, defaulting to stack,
but introducing new members in the CharacteristicsConfig struct will make
the source incompatible with targets other than SoftDevice.