linux: fix io.Reader impl for characteristics
The DeviceCharacteristics type implements io.Reader, but previously did
not follow the semantics as defined in the docs.
Several bugs are present that were fixed by this pull request:

1. the previous code always returned the full length of the data,
   regardless of how much data was actually read from the copy
2. io.EOF was never returned, meaning that we could never tell if the
   read was complete
3. the read always filled the buffer, then started over, meaning if the
   destination buffer was too small you would always get a partial read
   with no way to tell that the read was incomplete

This patch fixes this by caching the read internally, then copying from
the read slice. The value is not actually read again using the upstream
methods until the internal cached value is fully read.

Signed-off-by: Sam Whited <sam@samwhited.com>