Implement encoding interfaces for MAC address
Implements the encoding TextMarshaler, BinaryMarshaler, TextAppender and BinaryAppender interfaces for MAC for efficient encoding of MAC addresses.
The String method is written in terms of the new methods for consistency and improved performance.  This avoids unnecessary string allocations and conversions during the formatting process.
This also allows MAC type to be used with other encoding packages like encoding/json, encoding/xml.