Implement encoding interfaces for MAC address (#341)
* Add basic tests and benchmarks covering the mac.go file
* Implement encoding interfaces for MAC address
Implements the encoding TextMarshaler, BinaryMarshaler, TextAppender and BinaryAppender interfaces for MAC for efficient encoding of MAC addresses.
The String method is written in terms of the new methods for consistency and improved performance.  This avoids unnecessary string allocations and conversions during the formatting process.
This also allows MAC type to be used with other encoding packages like encoding/json, encoding/xml.

* Refactored while writing tests.

* Add unsafe string cast that matches the spirit of what uuid is doing.

The uuid package is using the string builder package.  That package
works in bytes and self manages length.  Because mac address string
length is always know this commit uses the same method as string builder
when casting the bytes to string.

* Add tests that cover MarshalText, AppendText, UnmarshalBinary and AppendBinary